+++ pc/src/table/ObstacleProximite.java
@@ -11,7 +11,7 @@ class ObstacleProximite extends ObstacleCirculaire {
 
 	public long death_date;
 	
-	public ObstacleProximite (Vec2 position, float rad, long death_date)
+	public ObstacleProximite (Vec2 position, int rad, long death_date)
 	{
 		super(position,rad);
 		this.death_date = death_date;
